Roles and permission refactoring
