+++ config/services.yaml
@@ -40,9 +40,11 @@ services:
         public: false
 
     PiaApi\Form\Type\RolesType:
-        arguments: ['@security.token_storage', '%security.role_hierarchy.roles%', '@security.role_hierarchy']
         tags: [form.type]
 
     PiaApi\Security\Voter\PermissionVoter:
         arguments: ['@security.role_hierarchy']
         tags: [security.voter]
+
+    PiaApi\Security\Role\RoleHierarchy:
+        arguments: ['@security.token_storage', '%security.role_hierarchy.roles%', '@security.role_hierarchy']
