+++ src/Controller/BackOffice/StructureController.php
@@ -12,7 +12,7 @@ namespace PiaApi\Controller\BackOffice;
 
 use Symfony\Component\HttpFoundation\Request;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use PiaApi\Entity\Pia\Structure;
 use PiaApi\Entity\Pia\Folder;
@@ -29,15 +29,10 @@ class StructureController extends BackOfficeAbstractController
 {
     /**
      * @Route("/manageStructures", name="manage_structures")
+     * @Security("is_granted('CAN_SHOW_STRUCTURE')")
      */
     public function manageStructuresAction(Request $request)
     {
-        if (!$this->isGranted('IS_AUTHENTICATED_REMEMBERED')) {
-            return $this->redirect($this->generateUrl('fos_user_security_login'));
-        }
-
-        $this->canAccess();
-
         $pagerfanta = $this->buildPager($request, Structure::class);
         $pagerfantaSt = $this->buildPager($request, StructureType::class, 20, 'pageSt', 'limitSt');
 
@@ -49,11 +44,10 @@ class StructureController extends BackOfficeAbstractController
 
     /**
      * @Route("/showStructure/{structureId}", name="manage_structures_show_structure")
+     * @Security("is_granted('CAN_SHOW_STRUCTURE')")
      */
     public function showStructureAction(Request $request)
     {
-        $this->canAccess();
-
         $structureId = $request->get('structureId');
         $structure = $this->getDoctrine()->getRepository(Structure::class)->find($structureId);
 
@@ -77,13 +71,12 @@ class StructureController extends BackOfficeAbstractController
 
     /**
      * @Route("/manageStructures/addStructure", name="manage_structures_add_structure")
+     * @Security("is_granted('CAN_CREATE_STRUCTURE')")
      *
      * @param Request $request
      */
     public function addStructureAction(Request $request)
     {
-        $this->canAccess();
-
         $form = $this->createForm(CreateStructureForm::class, [], [
             'action' => $this->generateUrl('manage_structures_add_structure'),
         ]);
@@ -112,13 +105,12 @@ class StructureController extends BackOfficeAbstractController
 
     /**
      * @Route("/manageStructures/addStructureType", name="manage_structures_add_structure_type")
+     * @Security("is_granted('CAN_CREATE_STRUCTURE_TYPE')")
      *
      * @param Request $request
      */
     public function addStructureTypeAction(Request $request)
     {
-        $this->canAccess();
-
         $form = $this->createForm(CreateStructureTypeForm::class, [], [
             'action' => $this->generateUrl('manage_structures_add_structure_type'),
         ]);
@@ -143,13 +135,12 @@ class StructureController extends BackOfficeAbstractController
 
     /**
      * @Route("/manageStructures/editStructure/{structureId}", name="manage_structures_edit_structure")
+     * @Security("is_granted('CAN_EDIT_STRUCTURE')")
      *
      * @param Request $request
      */
     public function editStructureAction(Request $request)
     {
-        $this->canAccess();
-
         $structureId = $request->get('structureId');
         $structure = $this->getDoctrine()->getRepository(Structure::class)->find($structureId);
 
@@ -179,13 +170,12 @@ class StructureController extends BackOfficeAbstractController
 
     /**
      * @Route("/manageStructures/editStructureType/{structureTypeId}", name="manage_structures_edit_structure_type")
+     * @Security("is_granted('CAN_EDIT_STRUCTURE_TYPE')")
      *
      * @param Request $request
      */
     public function editStructureTypeAction(Request $request)
     {
-        $this->canAccess();
-
         $structureTypeId = $request->get('structureTypeId');
         $structureType = $this->getDoctrine()->getRepository(StructureType::class)->find($structureTypeId);
 
@@ -215,13 +205,12 @@ class StructureController extends BackOfficeAbstractController
 
     /**
      * @Route("/manageStructures/removeStructure/{structureId}", name="manage_structures_remove_structure")
+     * @Security("is_granted('CAN_DELETE_STRUCTURE')")
      *
      * @param Request $request
      */
     public function removeStructureAction(Request $request)
     {
-        $this->canAccess();
-
         $structureId = $request->get('structureId');
         $structure = $this->getDoctrine()->getRepository(Structure::class)->find($structureId);
 
@@ -255,13 +244,12 @@ class StructureController extends BackOfficeAbstractController
 
     /**
      * @Route("/manageStructures/removeStructureType/{structureTypeId}", name="manage_structures_remove_structure_type")
+     * @Security("is_granted('CAN_DELETE_STRUCTURE_TYPE')")
      *
      * @param Request $request
      */
     public function removeStructureTypeAction(Request $request)
     {
-        $this->canAccess();
-
         $structureTypeId = $request->get('structureTypeId');
         $structureType = $this->getDoctrine()->getRepository(StructureType::class)->find($structureTypeId);
 
@@ -291,12 +279,5 @@ class StructureController extends BackOfficeAbstractController
         return $this->render('pia/Structure/removeStructureType.html.twig', [
             'form' => $form->createView(),
         ]);
-    }
-
-    protected function canAccess()
-    {
-        if (!$this->isGranted('ROLE_ADMIN')) {
-            throw new AccessDeniedHttpException();
-        }
     }
 }
