+++ src/Controller/Pia/AttachmentController.php
@@ -13,13 +13,14 @@ namespace PiaApi\Controller\Pia;
 use Symfony\Component\HttpFoundation\Request;
 use FOS\RestBundle\View\View;
 use FOS\RestBundle\Controller\Annotations as FOSRest;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use PiaApi\Entity\Pia\Attachment;
 
 class AttachmentController extends PiaSubController
 {
-  
     /**
      * @FOSRest\Get("/pias/{piaId}/attachments")
+     * @Security("is_granted('CAN_SHOW_PIA')")
      */
     public function listAction(Request $request, $piaId)
     {
@@ -28,6 +29,7 @@ class AttachmentController extends PiaSubController
 
     /**
      * @FOSRest\Get("/pias/{piaId}/attachments/{id}")
+     * @Security("is_granted('CAN_SHOW_PIA')")
      */
     public function showAction(Request $request, $piaId, $id)
     {
@@ -36,6 +38,7 @@ class AttachmentController extends PiaSubController
 
     /**
      * @FOSRest\Post("/pias/{piaId}/attachments")
+     * @Security("is_granted('CAN_EDIT_PIA')")
      */
     public function createAction(Request $request, $piaId)
     {
@@ -46,6 +49,7 @@ class AttachmentController extends PiaSubController
      * @FOSRest\Put("/pias/{piaId}/attachments/{id}")
      * @FOSRest\Patch("/pias/{piaId}/attachments/{id}")
      * @FOSRest\Post("/pias/{piaId}/attachments/{id}")
+     * @Security("is_granted('CAN_EDIT_PIA')")
      */
     public function updateAction(Request $request, $piaId, $id)
     {
@@ -54,6 +58,7 @@ class AttachmentController extends PiaSubController
 
     /**
      * @FOSRest\Delete("pias/{piaId}/attachments/{id}")
+     * @Security("is_granted('CAN_EDIT_PIA')")
      *
      * @return array
      */
