+++ src/Controller/Pia/CommentController.php
@@ -1,4 +1,5 @@
 <?php
+
 /*
  * Copyright (C) 2015-2018 Libre Informatique
  *
@@ -12,13 +13,14 @@ namespace PiaApi\Controller\Pia;
 use Symfony\Component\HttpFoundation\Request;
 use FOS\RestBundle\View\View;
 use FOS\RestBundle\Controller\Annotations as FOSRest;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use PiaApi\Entity\Pia\Comment;
 
 class CommentController extends PiaSubController
 {
-
     /**
      * @FOSRest\Get("/pias/{piaId}/comments")
+     * @Security("is_granted('CAN_SHOW_COMMENT')")
      */
     public function listAction(Request $request, $piaId)
     {
@@ -27,6 +29,7 @@ class CommentController extends PiaSubController
 
     /**
      * @FOSRest\Get("/pias/{piaId}/comments/{id}")
+     * @Security("is_granted('CAN_SHOW_COMMENT')")
      */
     public function showAction(Request $request, $piaId, $id)
     {
@@ -35,6 +38,7 @@ class CommentController extends PiaSubController
 
     /**
      * @FOSRest\Post("/pias/{piaId}/comments")
+     * @Security("is_granted('CAN_CREATE_COMMENT')")
      */
     public function createAction(Request $request, $piaId)
     {
@@ -45,6 +49,7 @@ class CommentController extends PiaSubController
      * @FOSRest\Put("/pias/{piaId}/comments/{id}")
      * @FOSRest\Patch("/pias/{piaId}/comments/{id}")
      * @FOSRest\Post("/pias/{piaId}/comments/{id}")
+     * @Security("is_granted('CAN_EDIT_COMMENT')")
      */
     public function updateAction(Request $request, $piaId, $id)
     {
@@ -53,6 +58,7 @@ class CommentController extends PiaSubController
 
     /**
      * @FOSRest\Delete("pias/{piaId}/comments/{id}")
+     * @Security("is_granted('CAN_DELETE_COMMENT')")
      *
      * @return array
      */
